Move kopt algorithms to a separate module
1. Move two k-opt functions to kopt.py
2. Remove k-opt functions from utils.py
